Add support for new output shape: 'impulse'.
In addition to continuing to support sin (by default now), this commit
introduces the idea of the output being of a given 'shape', which
currently is either 'sin' or 'impulse'.

Impulse has the additional parameter 'load-factor', which determines
the amount of time that is spent being 'on' during an interval. This
was introduced in an attempt to measure the room response.